+++ main_scene.go
@@ -77,14 +77,14 @@ func (s *MainScene) Draw(d *Doodle) error {
 	})
 	label.Compute(d.Engine)
 	label.MoveTo(render.Point{
-		X: (d.width / 2) - (label.Size().W / 2),
+		X: (int32(d.width) / 2) - (label.Size().W / 2),
 		Y: 120,
 	})
 	label.Present(d.Engine, label.Point())
 
 	s.frame.Compute(d.Engine)
 	s.frame.MoveTo(render.Point{
-		X: (d.width / 2) - (s.frame.Size().W / 2),
+		X: (int32(d.width) / 2) - (s.frame.Size().W / 2),
 		Y: 200,
 	})
 	s.frame.Present(d.Engine, s.frame.Point())
